# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CodeInterview API
  description: |
    REST API for the Online Coding Interview Platform.
    
    This API enables real-time collaborative coding interviews with features including:
    - Session management (create, join, update, delete)
    - Participant management (add, remove, update status)
    - Code collaboration (save code snapshots, language switching)
    - Real-time updates via WebSocket
    
    **Note**: Real-time collaboration events (code changes, user presence) should be 
    implemented using WebSocket connections in addition to these REST endpoints.
  version: 1.0.0
  contact:
    name: API Support
    email: support@codeinterview.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.codeinterview.com/v1
    description: Production server
...
paths:
...
  /sessions/{sessionId}/participants:
    get:
      tags:
        - Participants
      summary: Get session participants
      description: Retrieve all participants in a session
      operationId: getParticipants
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Participants retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  participants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Participant'
  /sessions/{sessionId}/participants/{participantId}:
    delete:
      tags:
        - Participants
      summary: Remove participant
      description: Remove a participant from the session
      operationId: removeParticipant
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/ParticipantId'
      responses:
        '204':
          description: Participant removed successfully
        '404':
          description: Session or participant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Participants
      summary: Update participant
      description: Update participant information (e.g., online status)
      operationId: updateParticipant
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/ParticipantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParticipantRequest'
      responses:
        '200':
          description: Participant updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '404':
          description: Session or participant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Unique session identifier (8-character hex string)
      schema:
        type: string
        pattern: '^[a-f0-9]{8}$'
        example: 'abc12345'
    ParticipantId:
      name: participantId
      in: path
      required: true
      description: Unique participant identifier (UUID)
      schema:
        type: string
        format: uuid
        example: '550e8400-e29b-41d4-a716-446655440000'
  schemas:
...
    UpdateParticipantRequest:
      type: object
      properties:
        isOnline:
          type: boolean
          description: Participant online status
...
    Participant:
      type: object
      required:
        - id
        - name
        - role
        - joinedAt
        - isOnline
      properties:
        id:
          type: string
          format: uuid
          description: Unique participant identifier
        name:
          type: string
          description: Participant display name
        role:
          type: string
          enum: [interviewer, candidate]
          description: Participant role in the interview
        color:
          type: string
          description: Participant avatar color
        joinedAt:
          type: integer
          format: int64
          description: Timestamp when participant joined (milliseconds)
        isOnline:
          type: boolean
          description: Whether participant is currently online
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: 'Session not found'
        code:
          type: string
          description: Error code for programmatic handling
          example: 'SESSION_NOT_FOUND'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for authentication (optional for MVP).
        Future implementation should include user authentication.

# Note: Security is not enforced in MVP but should be added for production
# security:
#   - BearerAuth: []